refactor with index
