+++ client/src/components/QAModule/QAComponents/index.jsx
@@ -1,0 +1,4 @@
+export { default as AnswerEntry } from './AnswerEntry';
+export { default as QuestionEntry } from './QuestionEntry';
+export { default as QuestionsList } from './QuestionsList';
+export { default as SearchBar } from './SearchBar';
