+++ client/src/components/QAModule/index.jsx
@@ -1,0 +1,2 @@
+/* eslint-disable no-restricted-exports */
+export { default } from './QuestionsAndAnswers';
